Rework the workflow engine
This is a major rework of the workflow engine (i.e. Activiti
replacement).

The primary goal was to avoid race conditions by executing all
changes in a single repo MODIFY operation while checking the
"CaseType version not changed" precondition.

The secondary goal was to make the code more modular and therefore
maintainable. See Action and Request class trees.