Fix multiple activation mappings (MID-3419)
The condition "activation mapping should produce at least one value"
was checked for each mapping individually, leading to failures if
there were more of them; and some of them skipped because of true->false
condition transition.

NOTE: This commit also changes rule for weak mapping evaluation. This
is yet to be revisited; probably in the context of overall mapping code
cleanup (MID-3847).