Fix assignment policy plus/minus/zero evaluation
Assignment policy rules were evaluated using membership in plus, minus,
and zero set of "evaluated assignment triple", e.g. assignment was
considered to be added when it was present in the plus set. This was
not quite correct if validity come into play.

So here we introduced more precise AssignmentOrigin information and
its isBeingAdded, isBeingDeleted and isBeingModified methods. They
are used mainly for AssignmentModificationConstraintEvaluator and
HasAssignmentConstraintEvaluator.

This fixes MID-4944 and MID-4216.