Fix MID-6129 preliminarily
The item.diff method used in InboundProcessor is "too clever"
for the purpose. We do not want to do optimized, general comparison
of two items there. (Just by the way, this method does not provide it
anyway -- therefore the exception.)

We need simplified item.diffValues instead.

This is only a preliminary, quick hack. More serious implementation
will follow.