Refactor RepositoryCache code deeply
Breaking up RepositoryCache monolith to a set of cooperating classes.
It should be much more maintainable now. As part of this work
some minor issues were found and fixed.

This is a preparation for fixing of MID-6250.

Other changes:
1. Renaming CacheCounterManager to CounterManagerImpl.
2. Moving things not related to repository cache (CacheRegistry,
CacheDispatcherImpl, CounterManagerImpl) to separate packages.
3. Fixing provisioning.getObject (shadow, readonly) - it tried
to modify immutable object when applying definitions. Maybe other
similar issues still exist. They should be fixed when found.