Escape invalid XML chars in traces/audit
Sometimes an invalid XML character (like 0x06) is to be stored into
XML element content. This occurs especially when dealing with traces
or audit records. This commit fixes this: For traces it always
replaces such chars with a descriptive form. For audit records
this behavior is configurable with the default being turned off.

Resolves MID-6262.