Add support for provenance metadata
1) Built-in mapping for provenance (acquisition and yield) was
created. However, it is only a preliminary version. It cannot delete
values. This will require deeper changes in the framework.

2) Added support for metadata change on values KEPT during
consolidation. It is implemented as ADD & DELETE of the same value,
with different metadata. (Requires a lot of code changes.)

3) Prism equivalence/narrow methods were cleaned up a bit. This is
to enable correct diff'ing of values with metadata. This change has
a potential to break many things in midPoint.

4) Delta narrow method checks for ADD & DELETE of the same value.
Mainly to enable point 2 above.

5) Changed handling of deltas in the repository (to enable point 2).
This has broken various things and should be revisited.

6) Added consideringValueMetadata option in
ParameterizedEquivalenceStrategy.

7) Displaying basic value metadata in various debugDump/shortDump/etc
methods.

Related to MID-6275.