Restructure search-iterative task processing
In order to unify error handling and state/progress/error reporting
in search-iterative tasks (including reconciliation), the
search-iterative mechanism was restructured.

Instead of the traditional couple of task handler + result handler,
now we have four cooperating sets of classes:
- task handler,
- task execution,
- task part executions (see the "new" conceptual model of tasks),
- result handlers.

Also, further steps to streamline error handling in tasks were taken.
TaskException class was introduced. Error handling was extracted into
TaskExceptionHandlingUtil class. (Preliminarily.)

Other (minor) changes:
- Disabled "checkSubresult" flag in tests (will be removed).
- ScriptExecutionException has become a CommonException.
- ImportAccountsFromResourceTaskHandler is now
ImportFromResourceTaskHandler.
- SynchronizeAccountResultHandler functionality was split
into Synchronizer and auxiliary classes in import/reconciliation task.
- ReconciliationTaskResult.shadowReconCount now includes also protected
accounts (to be consistent with the other counters). It is used only
in the tests, anyway.
- Task handlers were moved to more appropriate packages.
- An option for doing separate objects/assignments validity
scanning within a single physical task was introduced.

This commit is a work in progress. All the tests seem to pass
but the code is not complete. In particular, existing code in
base classes was not cleaned up yet.