Streamline resource object classification
The long-term goal is to implement resource object classification
(i.e. determination of its kind/intent/tag) into provisioning.
This is a step towards this goal: although the code still resides
in model-impl, it is invoked by dedicated ResourceObjectClassifier
interface with clear semantics.

This allowed us to do the classification as part of shadow acquisition,
so the process is unified a little bit more now.

The unrelated change property was deleted.