Change isEmpty semantics for deltas + others
1) Item delta is now considered empty even if it has valuesToAdd
and valuesToDelete empty collections. Previous implementation required
them to be null. It's natural to consider a delta being empty also
in these cases; as it's application is a no-op.

2) "Add" object delta is now considered non-empty even if the object
being added is empty itself. The reason is that even such delta has
visible effect, and should not be ignored as being empty.

3) ObjectDelta is now marked as Cloneable (wasn't probably by mistake).

4) Added utility methods CloneUtil#cloneCloneable and
CloneUtil#cloneMembersToCollection.