PrettyPrinter: catch around toString call, returned exception message
With profiling level Arguments, toString() for Spring-enhanced services
throws IllegalAccessException (JDK 17) with "module java.base does not
open java.lang to unnamed module...".
This seems to be a Spring reflection problem but we don't want to suffer
from it - it would not affect profiling only, but it would fail the
whole operation as well.