Fix Definition#getTypeClass method
This method returned the static class for the given item or type.
But especially for properties it had thrown an exception if such class
would not exist (see XsdTypeMapper.toJavaType). This was not logical,
as for other cases this method happily returns null if the class was
not known.

Furthermore, there existed a method Definition#getTypeClassIfKnown that
fixed the problem with missing class for properties, but failed to be
correctly implemented for some other types of objects (e.g. containers).

Therefore, here we fix getTypeClass() to use
XsdTypeMapper.toJavaTypeIfKnown, so it no longer fails on unknown java
types. And we deleted getTypeClassIfKnown().