MID-8928: Introduced searchable annotation to schema
Finally we got to point, when searchable annotation is needed not only
for documentation purposes, but also filter parsing.

In order to support 8928 we need special handling for prism properties
with complex content (this may be decomposed and indexed in repository).
So axiom query parser will allow matches filter for prism properties
with complex type values, when marked as searchable.
 - Allowing this for all prism properties with complex values would result
   to posssible parsed filters, which are not executable and hard to debug.