Decouple normalization from matching rules
For a long time, the idea of normalization was tightly coupled with
the concept of matching rule. For example, distinguishedName matching
rule involved the normalization algorithm for LDAP Distinguished names.

This was now changed. A normalizer prescribes the normalization
and matching algorithm, i.e. how a value is normalized, and how
two values are matched. This can be applied to polystring creation, by
specifying how the "norm" values should be derived from "orig" ones.
But for the actual querying that involves polystrings, two distinct
matching rules are to be used: polyStringOrig (matching original value)
and polyStringNorm (matching the normalized value).

Work in progress. This approach is not definite. The code is not
complete.

We may consider replacing the term "matchingRule" by "normalization"
for attribute definitions.