Support PolyString attributes in the repository
This commit brings the necessary support for the PolyString repo shadow
attributes:

- It removes the assumption that there is a single PolyString
normalization algorithm in the system. So, for example,
PolyStrings are not recomputed if there is already a norm value there.

- There are minor changes around matching rules, normalizers,
diagnostics. In particular, the real PolyStringOrig/Norm normalizers are
finally provided that comply with the definition of Normalizer<T>.

- There are minor changes in parsing algorithm needed to support
xsi:type for shadow attributes (see
SchemaRegistryImpl#locateItemDefinition).

- Application and re-application of definitions to prism
items and values was clarified a bit. (But a set of PolyString/String
hacks were introduced to PrismPropertyValueImpl, to be resolved later.)

Related to MID-2119 (shadow caching).