Support native midPoint associations
Main goal
---------

To support native midPoint associations. It required massive
changes in handling native (formerly raw) resource schema support
in midPoint, cascading down here to prism.

Main change
-----------

Decoupled schema parsing and serialization from prism definitions
(derived from Definition interface). The parsing and serialization
now use more generic/focused interfaces, like SchemaBuilder,
DefinitionBuilder, TypeDefinitionLikeBuilder, and so on (for parsing),
and SerializableSchema, SerializableDefinition, etc (for serialization);
because we needed to parse and serialize native resource definitions,
which are no longer prism ones.

EXPERIMENTAL. To be seen if it was a good idea at all. But it helped us
to move forward with the work on shadow associations.

Specific changes:

A. Schema parsing and serialization
-----------------------------------

1. Schema XSD parsing and serialization was largely reworked. Besides
aforementioned decoupling from prism definitions, they now use
DefinitionFeature objects to be more modular and understandable.

2. Existing Annotation and AnnotationProcessor classes are now
implemented using DefinitionFeature and DefinitionFeatureParser classes.
Behavioral change: annotations are no longer stored into the dynamic
"annotations" map of the definition. To be reviewed/resolved later.

3. The access to schema parsing/serialization from upper layers was
simplified. Instead of SchemaDefinitionFactory and its extension
we use a declarative list of DefinitionFeature instances to parse or
serialize for given complex type or item.

4. Fixed the issue with putting unrelated definitions into a schema.
(Occurred when parsing a schema that referenced types from other
schemas. We now parse a set of related schemas at the same time.)

5. Naming changes: DomToSchema processor renamed to SchemaDomParser.
DomToSchemaPostProcessor renamed to SchemaXsomParser.
SchemaToDomProcessor renamed to SchemaDomSerializer.

B. Definition classes
---------------------

1. Definition, ItemDefinition, TypeDefinition and some of their children
had specific aspects factored out into separate types, like
PrismItemBasicDefinition, PrismLifecycleDefinition, and so on. This
should make the types more manageable. The primary reason is to allow
creation of non-prism definitions like for native resource attributes.

2. Changes in definitions:
- "Processing" was moved from Definition to ItemDefinition.
It has a meaning for specific item only, anyway.
- "Indexed" was moved from property definition to item definition;
it sits now well in PrismItemStorageDefinition.

2. MutableXXX classes were converted to inner classes of their
immutable counterparts. What is more important, though, that they no
longer extend them: so e.g. MutableItemDefinition no longer extends
ItemDefinition. It can be used solely for changing values.

3. Type name and item name are now immutable in definitions.
This is because these are sometimes used as keys in maps, and form
a kind of identity for the definition.

4. Removed some unused/less-used methods in the definition interfaces:
canBeDefinitionOf, adoptElementDefinitionFrom.

5. Instantiations of definition objects (like ComplexTypeDefinitionImpl)
was a bit streamlined. Many of the constructors are no longer public.
DefinitionFactory should be used instead.

C. Other
--------

1. Removed dereferencing object getter/setters like #getLink / #setLink
(meaning getLinkRef.getObject and the analogous setter). These were not
used much anyway, and complicated the code needlessly.

2. Introduced "no access" enum for AccessAnnotationType. We need that
to represent totally hidden attributes - for the hidden side of native
associations.

3. Removed deprecated PrismContextSensitive interface. Prism context
is now statically accessed at all places.

Work in progress. Not polished enough. Parts of the new code needs to
be streamlined. Maybe the whole xyzDelegator and xyzMutatorDelegator
idea in definition fragments is wrong, and they should be simply
unrolled in the class where they are needed. But that would not be
a principal change. The key question is what is the optimal form
of schema parsing and serialization.