MID-9658: For boolean also generate get form of getter
Our code model always used form `is` for nullable boolean getters,
Groovy used to mapp this to properties.
But Groovy 4 does not map nullable `is` getters to to properties consistently anymore.

Fix is to generate also secondary getter for boolean in form `get`
which is consistently mapped to Groovy property.