Stop narrowing deltas when raw
If the delta or the base item is raw, the "narrow" method does not
work adequately (because of pitfalls related to comparison of raw
values). Perhaps the simplest and most effective solution is to
disable narrowing the delta if either the delta or the base item
itself has some raw parts.

This resolves MID-7918 (the part related to ignoring modifications
of "const" based configuration properties).

(cherry picked from commit 7c193d31d3b978a529badc4ad70e4447758673d8)