Implement removal of metadata from paths
**What**

Implement new `PrismContainerValue` method, which allows to delete
metadata from specified paths.

**Why**

This function is implemented mainly for the use by REST APIs, which
could exclude metadata based on user's request. Till now user could
exclude "normal" data, but not metadata.

**Note**

This change also bring new `Walkable` interface, which provide method to
walk through container items hierarchy. Implementation of this method is
used by the new method for metadata removal as well as for the existing
`removePaths` method.

Currently this new interface is extended only by the
`PrismContainerValue`. I have considered if it should be extended also
by the `PrismContainer` itself, but it would require to write
implementation in more places which I am not so far familiar enough. I
may add it later if I will be advised to do so.

**Fixes**: MID-10216