Apply filter definition on filter value
**What**

Apply filter definition (if any) to all filter values.

**Why**

It may happen (e.g. when defined by UI or script), that in time of
filter creation, the value passed to filter is not of the same type as
specified by the definition itself. For some types it is fine (e.g.
String value, PolyString in definition), because they can be converted.

In order to "force" this conversion, the definition is now applied to
the value on following circumstances:

- Filter is created with both value(s) and definition
- Definition is set to the filter after its creation (via setter)
- Value(s) is added to the filter after its creation (via setter)

The reason, why we want to convert the types using the definition is,
that value of this filter may be part of matching with value from
repository. Depending on the matcher and normalizer, it may happen that
during comparison of those two values, there will be thrown
`ClassCastException`, because the types are different and can't be
directly cast-ed.

**Note**

The application of definition to value may fail (e.g. when types are not
"convertible"). We decided to ignore this event (`SchemaException`) and
silently "swallow" it. The reason for this unfortunate thing is that (at
this moment) we don't want to change API contract and break backward
compatibility, because these filters can be used directly via user's
scripts.

**Fixes**: MID-10048