Fix working with frozen objects
Before this commit, if asPrismObject() was called on frozen Objectable,
the result was NOT immutable PrismObject. (But with immutable content.)
This made CloneUtil.toImmutable() inefficient, as it had to clone the
whole object.

This commit fixes that by preserving immutability when switching
between Objectable and PrismObject form.

It also removes ItemFactory#createObjectValue and createContainerValue
methods that were not used much (and complicated things related to
instantiation of prism objects).