Improve documentation of delta builder
**What**

Add an important note to one of the delta builder interfaces, about its
potential immutability.

**Why**

Currently the only implementation of this interface is `DeltaBuilder`
class, which implements affected methods in an "immutable" way, meaning
that they does not return the same instance of the builder, but creates
a new one.

That is however a bit confusing, because other methods does not behave
like this. If the returned instance is not correctly reassigned to the
initial variable, it can cause bugs where not all added items will be
present in built delta.

The docs uses the word "potential", because I do not want to add
behavior as a constraint to other implementers.